Attempt at enabling selection of multiple answers in polls
This is a failed attempt at extending the usability of polls by letting users vote for multiple
options, based on the max_selections parameter. While this attempt somewhat allows
for that, it is quite unusable as answers do not really persist through any selections made
after a page reload. I think this is because the new `selections` state gets reset to null
in these cases. I tried to address that by adding a `componentDidMount` call,
but that did not result in a notable improvement.
Besides the stated issue there are mainly two other missing features for this to work
in production: the ability for poll creators the manually set the max_selections parameter,
and the ability for voters to unselect made votes.